+++ CHANGELOG.md
@@ -6,6 +6,12 @@
     * Add `fundingSource` to the app switch url link. 
     * Add `funding_source` attribute to instrumentation events. 
 
+## 5.21.0 (2026-01-05)
+
+* UIComponents
+    * Add `PayPalButton` class to generate a PayPal-branded button for launching PayPal flow
+    * Add `VenmoButton` class to generate a Venmo-branded button for launching Venmo flow
+
 ## 5.20.0 (2025-12-17)
 
 * All Modules